refactor(types): promote NormalizeMAC plan modifier to shared types package
Move the set-of-MAC normalization plan modifier (and its unit test) out of
the apgroup domain into the shared internal/provider/types package so other
resources (e.g. the new unifi_setting_global_switch) can reuse it without a
cross-domain import. Behavior is unchanged; unifi_ap_group now calls
ut.NormalizeMAC().
